create a supernode that has nodeId of the entire network peers
we need to have supernode to store the result cache in the corresponding
node directly since we are not interested in the
steps of "storing the result cache" and we don't calculate anything
related to it.
we may also need it to observe some parameters in future. <br>
**future if we needed to calculate the bandwidth cost or time cost of
result cache storage, then we have to use proper DHT steps.